# docs/swagger.yaml
basePath: /
definitions:
...
  models.Correo:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Direccion:
    properties:
      area:
        type: string
      created_at:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      estado:
        type: string
      id:
        type: string
      lugar:
        type: string
      municipio:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
    type: object
...
  models.Empresa:
    properties:
      actividad_economica:
        type: string
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      id:
        type: string
      nombre:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      rif:
        type: string
      socios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.IIO:
    properties:
      area:
        type: string
      created_at:
        type: string
      descripcion:
        type: string
      fecha:
        type: string
      id:
        type: string
      imagen_url:
        type: string
      lugar:
        type: string
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      redi:
        type: string
      relacion:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tie:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      verificada:
        type: boolean
      zodi:
        type: string
    type: object
...
  models.Nacionalidad:
    properties:
      area:
        type: string
      created_at:
        type: string
      gentilicio:
        type: string
      id:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Pasaporte:
    properties:
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      foto:
        type: string
      id:
        type: string
      numero:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Persona:
    properties:
      alias:
        type: string
      apellido:
        type: string
      area:
        type: string
      cargo_actual:
        type: string
      cedula:
        type: string
      correo:
        type: string
      correos:
        items:
          $ref: '#/definitions/models.Correo'
        type: array
      created_at:
        type: string
      direcciones:
        items:
          $ref: '#/definitions/models.Direccion'
        type: array
      empresas:
        items:
          $ref: '#/definitions/models.Empresa'
        type: array
      estado_civil:
        type: string
      filiacion_politica:
        type: string
      id:
        type: string
      ideologia:
        type: string
      iio:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
      informacion_de_interes:
        type: string
      nacionalidad:
        items:
          $ref: '#/definitions/models.Nacionalidad'
        type: array
      nombre:
        type: string
      pasaporte:
        items:
          $ref: '#/definitions/models.Pasaporte'
        type: array
      profesion:
        type: string
      redes:
        items:
          $ref: '#/definitions/models.Redes'
        type: array
      relacionados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      religion:
        type: string
      telefono:
        type: string
      tipo_perfil:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoraciones:
        type: string
      vehiculos:
        items:
          $ref: '#/definitions/models.Vehiculo'
        type: array
      visa:
        items:
          $ref: '#/definitions/models.Visa'
        type: array
    type: object
  models.Redes:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Vehiculo:
    properties:
      area:
        type: string
      año:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      linea:
        type: string
      marca:
        type: string
      matricula:
        type: string
      modelo:
        type: string
      numero:
        type: string
      propietario:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      vehiculo:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
    type: object
  models.Visa:
    properties:
      aprobada:
        type: boolean
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      id:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoracion:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API SISTEMA DE EXPLORACION POPULAR.
  title: API
  version: "1.0"
paths:
...
  /pasaportes:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Número del pasaporte
        in: formData
        name: numero
        required: true
        type: string
      - description: Foto del pasaporte
        in: formData
        name: foto
        type: file
      - description: País del pasaporte
        in: formData
        name: pais
        required: true
        type: string
      - description: Tipo del pasaporte
        in: formData
        name: tipo
        required: true
        type: string
      - description: Código único del pasaporte
        in: formData
        name: codigo
        required: true
        type: string
      - description: ID del representante
        in: formData
        name: representante_id
        required: true
        type: string
      - description: Nivel del pasaporte
        in: formData
        name: nivel
        required: true
        type: string
      - description: ID del usuario
        in: formData
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Pasaporte'
      summary: Crea un nuevo pasaporte
      tags:
      - Pasaporte
  /pasaportes/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del pasaporte a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Pasaporte eliminado exitosamente
      summary: Elimina un pasaporte por su ID
      tags:
      - Pasaporte
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del pasaporte
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pasaporte'
      summary: Obtiene un pasaporte por su ID
      tags:
      - Pasaporte
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del pasaporte a actualizar
        in: path
        name: id
        required: true
        type: string
      - description: Número del pasaporte
        in: formData
        name: numero
        required: true
        type: string
      - description: Foto del pasaporte
        in: formData
        name: foto
        type: file
      - description: País del pasaporte
        in: formData
        name: pais
        required: true
        type: string
      - description: Tipo del pasaporte
        in: formData
        name: tipo
        required: true
        type: string
      - description: Código único del pasaporte
        in: formData
        name: codigo
        required: true
        type: string
      - description: ID del representante
        in: formData
        name: representante_id
        required: true
        type: string
      - description: Nivel del pasaporte
        in: formData
        name: nivel
        required: true
        type: string
      - description: ID del usuario
        in: formData
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pasaporte'
      summary: Actualiza un pasaporte existente por su ID
      tags:
      - Pasaporte
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"